requirements.txt
# Core dependencies
pyaudio==0.2.13
torch>=1.12.0
numpy>=1.20.0
torchaudio>=0.12.0
pyyaml>=6.0
rich>=13.0.0

# Speech recognition
faster-whisper>=0.9.0
ffmpeg-python>=0.2.0
ctranslate2>=3.14.0 